requirements.txt
discord.py==2.4.0
PyNaCl==1.5.0
python-dotenv==1.0.1
httpx==0.27.0
deep-translator==1.11.4
motor==3.7.1
pymongo==4.15.4
aiohttp>=3.8.5
beautifulsoup4==4.12.2
edge-tts==6.1.10
aiofiles